Move the plugin into test-plugins.js
